apply code-review feedback
