+++ lib/sampler/storage/src/lib/span-dao.js
@@ -150,6 +150,11 @@ class SpanDAO {
   }
 }
 
+// NOTE: Changing index definition in the DAO will result in failure
+// for existing deployments, changing options will be ignored.
+// Should that ever become necessary, either some type of update logic
+// needs to be implemented here, or an external provisioning step needs to perform it.
+
 const createTargetIndex = async (collection) => {
   const definition = {
     'target.organization_id': 1,
