+++ lib/sampler/storage/src/test/span-dao-test.js
@@ -5,88 +5,13 @@ const { extend } = require('underscore');
 const uuid = require('uuid');
 const moment = require('abacus-moment');
 const { createSpanDAO } = require('../lib/span-dao');
+const { createRandomTarget, spanPageObject } = require('./helpers/span-po');
 
 const mongoURI = process.env.DB || 'mongodb://localhost:27017';
 
-const createRandomTarget = () => ({
-  organization_id: uuid.v4(),
-  space_id: uuid.v4(),
-  consumer_id: uuid.v4(),
-  resource_id: uuid.v4(),
-  plan_id: uuid.v4(),
-  resource_instance_id: uuid.v4(),
-  correlation_id: uuid.v4()
-});
-
-// spanPageObject creates a page object that executes span commands
-// for a specific span target, allowing for validation reuse
-const spanPageObject = (dao, target) => {
-
-  const startSpan = (timestamp, measures, dedupID = undefined) => {
-    const result = async () => {
-      return await dao.startSpan(timestamp, target, measures, dedupID);
-    };
-    const verify = async () => {
-      const success = await result();
-      expect(success).to.equal(true);
-    };
-    return { result, verify };
-  };
-
-  const endSpan = (timestamp, newCorrelationID, dedupID = undefined) => {
-    const result = async () => {
-      return await dao.endSpan(timestamp, target, newCorrelationID, dedupID);
-    };
-    const verify = async () => {
-      const success = await result();
-      expect(success).to.equal(true);
-    };
-    return { result, verify };
-  };
-
-  const getSpan = () => {
-    const result = async () => {
-      return await dao.getSpanByTarget(target);
-    };
-    const verify = async () => {
-      const span = await result();
-      expect(span).not.to.equal(undefined);
-      return span;
-    };
-    return { result, verify };
-  };
-
-  const updateSpanPlannedInterval = (interval, version) => {
-    const result = async () => {
-      const span = await getSpan().verify();
-      return await dao.updateSpanPlannedInterval(span._id, interval, version);
-    };
-    const verify = async () => {
-      const success = await result();
-      expect(success).to.equal(true);
-    };
-    return { result, verify };
-  };
-
-  const updateSpanProcessedInterval = (interval, complete, version) => {
-    const result = async () => {
-      const span = await getSpan().verify();
-      return await dao.updateSpanProcessedInterval(span._id, interval, complete, version);
-    };
-    const verify = async () => {
-      const success = await result();
-      expect(success).to.equal(true);
-    };
-    return { result, verify };
-  };
-
-  return {
-    startSpan,
-    endSpan,
-    getSpan,
-    updateSpanPlannedInterval,
-    updateSpanProcessedInterval
-  };
+const dateToTimestamp = (year, month, day, hour, minute, second, millis) => {
+  // in moment.js the month starts from 0 (i.e. Jan = 0), hence the subtraction
+  return moment.utc([year, month - 1, day, hour, minute, second, millis]).valueOf();
 };
 
 describe('span-dao', () => {
@@ -405,7 +330,7 @@ describe('span-dao', () => {
       await firstSpanPO.startSpan(startTimestamp, measures).verify();
       const firstInterval = {
         start: startTimestamp,
-        end: 1534161600000 // 2018-08-13 12:00:00
+        end: dateToTimestamp(2018, 8, 13, 12, 0, 0, 0)
       };
       await firstSpanPO.updateSpanProcessedInterval(firstInterval, false, 1).verify();
 
@@ -414,7 +339,7 @@ describe('span-dao', () => {
       await secondSpanPO.startSpan(startTimestamp, measures).verify();
       const secondInterval = {
         start: startTimestamp,
-        end: 1534168800000 // 2018-08-13 14:00:00
+        end: dateToTimestamp(2018, 8, 13, 14, 0, 0, 0)
       };
       await secondSpanPO.updateSpanProcessedInterval(secondInterval, false, 1).verify();
 
@@ -423,44 +348,50 @@ describe('span-dao', () => {
       await thirdSpanPO.startSpan(startTimestamp, measures).verify();
       const thirdInterval = {
         start: startTimestamp,
-        end: 1534176000000 // 2018-08-13 16:00:00
+        end: dateToTimestamp(2018, 8, 13, 16, 0, 0, 0)
       };
       await thirdSpanPO.updateSpanProcessedInterval(thirdInterval, true, 1).verify();
     });
 
     it('returns only spans that are sampled earlier than specified', async () => {
-      const spans = await dao.findIncompleteSpans(1534161600001, 0, 3); // 2018-08-13 12:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 12, 0, 0, 1);
+      const spans = await dao.findIncompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(1);
       expect(spans[0].target).to.deep.equal(firstTarget);
     });
 
     it('returns multiple spans, when applicable', async () => {
-      const spans = await dao.findIncompleteSpans(1534168800001, 0, 3); // 2018-08-13 14:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 14, 0, 0, 1);
+      const spans = await dao.findIncompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(2);
       expect(spans[0].target).to.deep.equal(firstTarget);
       expect(spans[1].target).to.deep.equal(secondTarget);
     });
 
     it('returns only spans that are not yet complete', async () => {
-      const spans = await dao.findIncompleteSpans(1534176000001, 0, 3); // 2018-08-13 16:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 16, 0, 0, 1);
+      const spans = await dao.findIncompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(2);
       expect(spans[0].target).to.deep.equal(firstTarget);
       expect(spans[1].target).to.deep.equal(secondTarget);
     });
 
     it('returns no candiates, if there are no matching', async () => {
-      const spans = await dao.findIncompleteSpans(1534161600000, 0, 3); // 2018-08-13 12:00:00.000
+      const before = dateToTimestamp(2018, 8, 13, 12, 0, 0, 0);
+      const spans = await dao.findIncompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(0);
     });
 
     it('returns spans according to the limit setting', async () => {
-      const spans = await dao.findIncompleteSpans(1534168800001, 0, 1); // 2018-08-13 14:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 14, 0, 0, 1);
+      const spans = await dao.findIncompleteSpans(before, 0, 1);
       expect(spans.length).to.equal(1);
       expect(spans[0].target).to.deep.equal(firstTarget);
     });
 
     it('returns spans according to the offset setting', async () => {
-      const spans = await dao.findIncompleteSpans(1534168800001, 1, 3); // 2018-08-13 14:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 14, 0, 0, 1);
+      const spans = await dao.findIncompleteSpans(before, 1, 3);
       expect(spans.length).to.equal(1);
       expect(spans[0].target).to.deep.equal(secondTarget);
     });
@@ -477,7 +408,7 @@ describe('span-dao', () => {
       await firstSpanPO.startSpan(startTimestamp, measures).verify();
       const firstInterval = {
         start: startTimestamp,
-        end: 1534161600000 // 2018-08-13 12:00:00
+        end: dateToTimestamp(2018, 8, 13, 12, 0, 0, 0)
       };
       await firstSpanPO.updateSpanProcessedInterval(firstInterval, true, 1).verify();
 
@@ -486,7 +417,7 @@ describe('span-dao', () => {
       await secondSpanPO.startSpan(startTimestamp, measures).verify();
       const secondInterval = {
         start: startTimestamp,
-        end: 1534168800000 // 2018-08-13 14:00:00
+        end: dateToTimestamp(2018, 8, 13, 14, 0, 0, 0)
       };
       await secondSpanPO.updateSpanProcessedInterval(secondInterval, true, 1).verify();
 
@@ -495,44 +426,50 @@ describe('span-dao', () => {
       await thirdSpanPO.startSpan(startTimestamp, measures).verify();
       const thirdInterval = {
         start: startTimestamp,
-        end: 1534176000000 // 2018-08-13 16:00:00
+        end: dateToTimestamp(2018, 8, 13, 16, 0, 0, 0)
       };
       await thirdSpanPO.updateSpanProcessedInterval(thirdInterval, false, 1).verify();
     });
 
     it('returns only spans that are sampled earlier than specified', async () => {
-      const spans = await dao.findCompleteSpans(1534161600001, 0, 3); // 2018-08-13 12:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 12, 0, 0, 1);
+      const spans = await dao.findCompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(1);
       expect(spans[0].target).to.deep.equal(firstTarget);
     });
 
     it('returns multiple spans, when applicable', async () => {
-      const spans = await dao.findCompleteSpans(1534168800001, 0, 3); // 2018-08-13 14:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 14, 0, 0, 1);
+      const spans = await dao.findCompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(2);
       expect(spans[0].target).to.deep.equal(firstTarget);
       expect(spans[1].target).to.deep.equal(secondTarget);
     });
 
     it('returns only spans that are not yet complete', async () => {
-      const spans = await dao.findCompleteSpans(1534176000001, 0, 3); // 2018-08-13 16:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 16, 0, 0, 1);
+      const spans = await dao.findCompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(2);
       expect(spans[0].target).to.deep.equal(firstTarget);
       expect(spans[1].target).to.deep.equal(secondTarget);
     });
 
     it('returns no candiates, if there are no matching', async () => {
-      const spans = await dao.findCompleteSpans(1534161600000, 0, 3); // 2018-08-13 12:00:00.000
+      const before = dateToTimestamp(2018, 8, 13, 12, 0, 0, 0);
+      const spans = await dao.findCompleteSpans(before, 0, 3);
       expect(spans.length).to.equal(0);
     });
 
     it('returns spans according to the limit setting', async () => {
-      const spans = await dao.findCompleteSpans(1534168800001, 0, 1); // 2018-08-13 14:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 14, 0, 0, 1);
+      const spans = await dao.findCompleteSpans(before, 0, 1);
       expect(spans.length).to.equal(1);
       expect(spans[0].target).to.deep.equal(firstTarget);
     });
 
     it('returns spans according to the offset setting', async () => {
-      const spans = await dao.findCompleteSpans(1534168800001, 1, 3); // 2018-08-13 14:00:00.001
+      const before = dateToTimestamp(2018, 8, 13, 14, 0, 0, 1);
+      const spans = await dao.findCompleteSpans(before, 1, 3);
       expect(spans.length).to.equal(1);
       expect(spans[0].target).to.deep.equal(secondTarget);
     });
